Changes throughout
I got set back a little due to being dumb and using Bebas. I have it on
my computer but I overheard that that doesn’t fly around here. So I
searched for a new typeface and found Oswald. I had to re-adjust to
meet the needs of the new typeface I also started get my images set and
ready to edit in photoshop. Then other basic edits to my html and css.